extract the wizard form functionality
